Adding DiagnosticSource to EventSource Bridge.
Currently DiagnosticSource just goes to DiagnosticListeners.   We want to
be able to have the capabilty to pipe DiagnosticSource information to
EventSource consumers (ETW and EventListeners).  We do this with a new
DiagnosticSource EventSource called 'Microsoft-DiagnosticsSource'  If
EventListener / ETW subscribe to that EventSource they can get events
from any DiangnosticSouce.

See the comments in the code itself as well as the tests in the file
DiagnosticSourceEventSourceBridgeTests.cs for more on specifics of
using this new functionality.


Commit migrated from https://github.com/dotnet/corefx/commit/34f72a1ff07fa589544558127867dced52e710d8