Fix library packages overlapping with framework packages
Many library packages overlap with assemblies we include in our framework packages.

In the first release we handled this by preventing the library package from applying any
assets.

This becomes problematic if the reference assembly needs to version, and especially
if the library wants to add API.  The new version will be greater than what was previously
shipped inbox, and we wouldn't want to restrict the new version just to future TFMs.

To address this problem I've made a couple fixes:
1. For packages which are shipping brand new API that's not present in desktop, we'll
omit any placeholders and allow conflict resolution to decide if the library package
or framework package should be used (based on version).  In some cases this required
bringing some configurations into the package so that it could always be "up to date".

2. For packages which are just exposing API in desktop, and thus already tied/anchored
to netstandard versions, we'll freeze the reference assembly version for the non-desktop
builds.  This ensures that the reference assembly will never move past the version that's
inbox in the framework package.

Update BuildToolsVersion to 2.0.0-prerelease-01931-01


Commit migrated from https://github.com/dotnet/corefx/commit/7ee51be7ed472a12266c8cc2ab25a9e1862c1d00