Forking the DataDog .NET Tracer to seed the OTel .NET Auto-Instrumentation-Agent.
This is a fork of the DataDog .NET Tracer (aka "in-process agent", aka "auto-instrumentation agent").
It acts as the initial seed for the official Open Telemetry .NET Tracer.

We are doing this by merging the current master of https://github.com/DataDog/dd-trace-dotnet into an empty master of https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation.
In addition, I cleaned up all auxiliary files not directly related to the product (e.g., demos, tools, vendor-specific integration tests and benchmarks, we well as others). We can add them in later from the original repo if the community requires.

Everything builds, all unit tests are passing.