Modify our HttpModule to propagate the newly created span context through HTTP headers so the WCF pipeline can pick up the distributed trace context (#1018)
## Background
When a WCF service is hosted in IIS, the WCF service is hosted [side-by-side](https://docs.microsoft.com/en-us/dotnet/framework/wcf/feature-details/wcf-services-and-aspnet) with the ASP.NET pipeline so it cannot retrieve the current Datadog scope that is created by the HttpModule.

## Changes
Once the HttpModule creates an `aspnet` span, the HttpModule will also inject the trace context into the HTTP request via the distributed tracing headers. Then, when the WCF pipeline is later invoked, the trace context of the `aspnet` span will be extracted from the HTTP headers of the request object (unchanged in this PR), which then connects the spans into one trace.

## Testing
This has been validated in the following scenarios when hosting a WCF service on IIS:
1. HTTP request from browser (no distributed tracing)
2. Distributed trace

In both cases, the traces that were incorrectly created in parallel are now nested correctly.