Add configuration for customising span service names (#1142)
* Fix RabbitMQ not using correct servicename

* Add DD_TRACE_SERVICE_MAPPING configuration key

Used to rename a sub-service name based on configuration
For most integrations, the key is the integration type
For DbCommands, this is based on dbType.
For HttpClient, we map all http-clients, regardless of destination

* Add fastpath avoiding dictionary lookup when no service name mappings

* Add integration test for service mapping

* Wrap ServiceNameMappings dictionary to avoid direct external access

Additional mappings can be added using the `SetServiceNameMappings` method 
The default path, where there are no mappings, avoids the dictionary lookup. If required, we can further optimise this later (e.g. handle small number of mappings, known integration service names etc)

* Add support for null IDictionary in .ctor

Avoids creating an unneccessary instance in TracerSettings

* Use IsNullOrWhiteSpace when creating configuration key values

* Update service mappings API to allow replacing whole configuration