Make it the default behavior to pass the CallTargetState object by ref. This updates both the native and managed side to the new behavior and removes the need for the EnableCallTargetStateByRef PInvoke.
Note: In some parts of this commit, this commit decides to keep the existing `enable_calltarget_state_by_ref` conditional branches intact, to make it easier for future upstream updates. However, if we'd like to remove those branches entirely (since only one will ever be called), we can do so.